// app/routes/staff/types/staff-body/StaffBody.tsx
export interface StaffBody {
  id: string;
  first_name: string;
  middle_name: string | null;
  last_name: string;
  email: string;
  phone: string | null;
  position: string | null;
  department: string | null;
  gender: string | null;
  religion: string | null;
  marital_status: string | null;
  country: string | null;
  city: string | null;
  nationality: string | null;
  staff_status: string | null;
  hire_date: string | null;
  date_of_birth: string | null;
  address: string | null;
  emergency_contact_name: string | null;
  emergency_contact_relationship: string | null;
  emergency_contact_phone: string | null;
  next_of_kin_name: string | null;
  next_of_kin_relationship: string | null;
  next_of_kin_phone: string | null;
  id_type: string | null;
  id_number: string | null;
  id_issue_date: string | null;
  id_expiry_date: string | null;
  is_transfered: boolean;
  transfer_date: string | null;
  transfer_reason: string | null;
  work_email: string | null;
  work_phone: string | null;
  languages: string | null;
  academic_qualifications: string | null;
  professional_qualifications: string | null;
  skills: string | null;
  photo: string | null;
  created_at: string;
  updated_at: string;
}